4.4.6
Fixed the Creek error.
Changed a lind at the start of int_escape to fit a better narative.
Changed the subs for int_escape intro into something more fitting the world of GFL.
As it fits better with where 40 is from according to the game and point out of the cover up of OP butterfly.